Move more REST files into src (#37026)
