+++ components/DefaultLayout.tsx
@@ -7,7 +7,7 @@ import { LegalFooter } from 'components/page-footer/LegalFooter'
 import { ScrollButton } from 'components/ui/ScrollButton'
 import { SupportSection } from 'components/page-footer/SupportSection'
 import { DeprecationBanner } from 'components/page-header/DeprecationBanner'
-import { RestBanner } from 'components/page-header/RestBanner'
+import { RestBanner } from 'src/rest/components/RestBanner'
 import { useMainContext } from 'components/context/MainContext'
 import { useTranslation } from 'components/hooks/useTranslation'
 import { Breadcrumbs } from 'components/page-header/Breadcrumbs'
