+++ src/rest/components/RestCollapsibleSection.module.scss
