+++ src/rest/components/useClipboard.ts
